+++ src/types/common.d.ts
@@ -1,38 +1,3 @@
-//** TODO: Simplify FonData, is it used? */
-export interface FontData {
-  glyphs: { x_min: number; x_max: number; ha: number; o: string[] }[];
-  cssFontWeight: string;
-  ascender: number;
-  underlinePosition: number;
-  cssFontStyle: string;
-  boundingBox: { yMin: number; xMin: number; yMax: number; xMax: number };
-  resolution: number;
-  descender: number;
-  familyName: string;
-  lineHeight: number;
-  underlineThickness: number;
-  /**
-   * This property is missing in vexflow_font.js, but present in gonville_original.js and gonville_all.js.
-   */
-  original_font_information?: {
-    postscript_name: string;
-    version_string: string;
-    vendor_url: string;
-    full_font_name: string;
-    font_family_name: string;
-    copyright: string;
-    description: string;
-    trademark: string;
-    designer: string;
-    designer_url: string;
-    unique_font_identifier: string;
-    license_url: string;
-    license_description: string;
-    manufacturer_name: string;
-    font_sub_family_name: string;
-  };
-}
-
 export interface KeyProps {
   stem_down_x_offset: number;
   stem_up_x_offset: number;
